+++ src/assets/components/SubmitButton.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+export function SubmitButton({ onClick }) {
+  return (
+    <button className="submitButton" onClick={onClick}>
+      Submit
+    </button>
+  );
+}
+
+
