+++ src/index.css
@@ -17,68 +17,3 @@ code {
   padding: 0;
   box-sizing: border-box;
 }
-
-.header {
-  height: 350px;
-  background-color: #0018a4;
-  text-align: center;
-  font-family: "Poppins", sans-serif;
-  padding: 70px;
-  text-align: center;
-  padding: 100px 50px 60px 50px;
-  margin: auto;
-  line-height: 45px;
-}
-
-.page-title {
-  font-weight: 500;
-  font-size: 40px;
-  color: white;
-}
-
-.page-subtitle {
-  font-size: 30px;
-  font-weight: 400;
-  color: white;
-}
-
-.QuestionForm {
-  text-align: center;
-  padding: 80px;
-  font-family: "Poppins", sans-serif;
-  line-height: 60px;
-  color: #0018a4;
-  line-height: 45px;
-}
-
-.number {
-  font-size: 40px;
-  font-weight: 500;
-}
-
-.question {
-  font-size: 30px;
-  font-weight: 400;
-  margin: 20px;
-}
-
-.activity {
-  width: 500px;
-  height: 60px;
-  border-radius: 6px;
-  padding: 10px;
-}
-
-.RadioButtons {
-}
-
-/* input[type="radio"]{
-  display: none;
-} */
-
-label {
-  position: relative;
-  font-family: "Poppins", sans-serif;
-  border-radius: 5px;
-
-}
